pyproject.toml
[tool.poetry]
name = "education"
version = "1.0.0"
description = ""
authors = ["Alex Gusev <gusev@ibs.ru>"]


[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.109.2"
pydantic = "^2.6.1"
uvicorn = "^0.27.0.post1"
python-dotenv = "^1.0.0"
python-dateutil = "^2.8.2"
concurrent-log-handler = "^0.9.25"
sqlalchemy = "^2.0.25"
asyncpg = "^0.29.0"
pandas = "^2.2.0"
pyyaml = "^6.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
